+++ src/com/kstenschke/referencer/utils/UtilsEnvironment.java
@@ -27,10 +27,7 @@ import java.awt.*;
 
 public class UtilsEnvironment {
 
-    /**
-     * @return The currently opened project
-     */
-    private static Project getOpenProject() {
+    public static Project getOpenProject() {
         Project[] projects = ProjectManager.getInstance().getOpenProjects();
 
         return (projects.length > 0) ? projects[0] : null;